Move implementation out of problems api project
