+++ platforms/extensibility/plugin-development/src/main/java/org/gradle/plugin/devel/tasks/internal/ValidationProblemSerialization.java
@@ -41,7 +41,6 @@ import org.gradle.api.problems.internal.DefaultLineInFileLocation;
 import org.gradle.api.problems.internal.DefaultOffsetInFileLocation;
 import org.gradle.api.problems.internal.DefaultPluginIdLocation;
 import org.gradle.api.problems.internal.DefaultProblem;
-import org.gradle.api.problems.internal.DefaultProblemCategory;
 import org.gradle.api.problems.internal.DefaultProblemGroup;
 import org.gradle.api.problems.internal.DefaultProblemId;
 import org.gradle.api.problems.internal.DefaultPropertyTraceData;
@@ -54,7 +53,6 @@ import org.gradle.api.problems.internal.GeneralData;
 import org.gradle.api.problems.internal.LineInFileLocation;
 import org.gradle.api.problems.internal.OffsetInFileLocation;
 import org.gradle.api.problems.internal.Problem;
-import org.gradle.api.problems.internal.ProblemCategory;
 import org.gradle.api.problems.internal.ProblemLocation;
 import org.gradle.api.problems.internal.PropertyTraceData;
 import org.gradle.api.problems.internal.TypeValidationData;
@@ -65,7 +63,6 @@ import javax.annotation.Nullable;
 import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Type;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -88,7 +85,6 @@ public class ValidationProblemSerialization {
         gsonBuilder.registerTypeAdapter(ProblemId.class, new ProblemIdInstanceCreator());
         gsonBuilder.registerTypeHierarchyAdapter(DocLink.class, new DocLinkAdapter());
         gsonBuilder.registerTypeHierarchyAdapter(ProblemLocation.class, new LocationAdapter());
-        gsonBuilder.registerTypeHierarchyAdapter(ProblemCategory.class, new ProblemCategoryAdapter());
         gsonBuilder.registerTypeHierarchyAdapter(AdditionalData.class, new AdditionalDataAdapter());
         gsonBuilder.registerTypeAdapterFactory(new ThrowableAdapterFactory());
 
@@ -442,61 +438,6 @@ public class ValidationProblemSerialization {
                     return finalConsultDocumentationMessage;
                 }
             };
-        }
-    }
-
-    public static class ProblemCategoryAdapter extends TypeAdapter<ProblemCategory> {
-
-        @Override
-        public void write(JsonWriter out, @Nullable ProblemCategory value) throws IOException {
-            if (value == null) {
-                out.nullValue();
-                return;
-            }
-
-            out.beginObject();
-            out.name("namespace").value(value.getNamespace());
-            out.name("category").value(value.getCategory());
-            out.name("subcategories").beginArray();
-            for (String sc : value.getSubcategories()) {
-                out.value(sc);
-            }
-            out.endArray();
-            out.endObject();
-        }
-
-        @Override
-        public ProblemCategory read(JsonReader in) throws IOException {
-            in.beginObject();
-            String namespace = null;
-            String category = null;
-            List<String> subcategories = new ArrayList<>();
-            String name;
-            while (in.hasNext()) {
-                name = in.nextName();
-                switch (name) {
-                    case "namespace": {
-                        namespace = in.nextString();
-                        break;
-                    }
-                    case "category": {
-                        category = in.nextString();
-                        break;
-                    }
-                    case "subcategories": {
-                        in.beginArray();
-                        while (in.hasNext()) {
-                            subcategories.add(in.nextString());
-                        }
-                        in.endArray();
-                        break;
-                    }
-                    default:
-                        in.skipValue();
-                }
-            }
-            in.endObject();
-            return DefaultProblemCategory.create(namespace, category, subcategories.toArray(new String[0]));
         }
     }
