--- platforms/ide/problems-api/src/main/java/org/gradle/api/problems/internal/ProblemCategory.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.gradle.api.problems.internal;
-
-import java.util.List;
-
-public interface ProblemCategory {
-
-    /**
-     * The namespace.
-     *
-     * @return The namespace.
-     */
-    String getNamespace();
-
-    /**
-     * The main category.
-     *
-     * @return The main categories.
-     */
-    String getCategory();
-
-    /**
-     * The optional subcategories.
-     *
-     * @return The subcategories.
-     */
-    List<String> getSubcategories();
-}
