+++ platforms/ide/problems-api/src/main/java/org/gradle/api/problems/internal/ProblemsSummaryProgressDetails.java
@@ -19,5 +19,5 @@ package org.gradle.api.problems.internal;
 /*
  * Marker interface required by the build operations infrastructure.
  */
-interface ProblemsSummaryProgressDetails {
+public interface ProblemsSummaryProgressDetails {
 }
