+++ platforms/ide/problems/src/main/java/org/gradle/problems/internal/emitters/BuildOperationBasedProblemEmitter.java
@@ -17,11 +17,11 @@
 package org.gradle.problems.internal.emitters;
 
 import org.gradle.api.Incubating;
-import org.gradle.api.problems.internal.DefaultProblemProgressDetails;
 import org.gradle.api.problems.internal.Problem;
 import org.gradle.api.problems.internal.ProblemEmitter;
 import org.gradle.internal.operations.BuildOperationProgressEventEmitter;
 import org.gradle.internal.operations.OperationIdentifier;
+import org.gradle.problems.internal.events.DefaultProblemProgressDetails;
 
 import javax.annotation.Nullable;
