+++ platforms/ide/problems/src/main/java/org/gradle/problems/internal/events/DefaultProblemProgressDetails.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,13 @@
  * limitations under the License.
  */
 
-package org.gradle.api.problems.internal;
+package org.gradle.problems.internal.events;
 
+import org.gradle.api.NonNullApi;
+import org.gradle.api.problems.internal.Problem;
+import org.gradle.api.problems.internal.ProblemProgressDetails;
+
+@NonNullApi
 public class DefaultProblemProgressDetails implements ProblemProgressDetails {
     private final Problem problem;
