+++ subprojects/core/src/testFixtures/groovy/org/gradle/util/TestUtil.groovy
@@ -44,7 +44,6 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.api.problems.Problems
 import org.gradle.api.problems.internal.DefaultProblems
 import org.gradle.api.problems.internal.ExceptionProblemRegistry
-import org.gradle.api.problems.internal.NoOpProblemSummarizer
 import org.gradle.api.provider.ProviderFactory
 import org.gradle.api.tasks.util.internal.PatternSets
 import org.gradle.cache.internal.TestCrossBuildInMemoryCacheFactory
