+++ testing/internal-integ-testing/build.gradle.kts
@@ -97,6 +97,7 @@ dependencies {
     implementation(projects.messaging)
     implementation(projects.modelCore)
     implementation(projects.platformJvm)
+    implementation(projects.problems)
     implementation(projects.serialization)
     implementation(projects.serviceProvider)
     implementation(projects.serviceRegistryBuilder)
